[패키지리팩토링중..] util,errorhandler 기타등등 추가 후 html,js,css제거

+++ src/main/java/com/springboot/board/api/v1/dto/response/AnswerResponse.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class AnswerResponse {
-    private Long id; // 답변의 고유 ID
+    private Integer id; // 답변의 고유 ID
     private String content; // 답변 내용
-    // 필요한 필드 추가
+    
 }
