+++ src/main/java/com/springboot/board/domain/repository/AnswerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.springboot.board.domain.entity.Answer;
 
 @Repository
-public interface AnswerRepository extends JpaRepository<Answer, Long> {
+public interface AnswerRepository extends JpaRepository<Answer, Integer> {
 }
